Move ui tests into template/ folder
