+++ maple-core-macro/tests/template/component-fail.stderr
