+++ maple-core-macro/tests/template/root-pass.rs
