+++ maple-core-macro/tests/template_macro.rs
@@ -1,7 +1,7 @@
 #[test]
 fn ui() {
     let t = trybuild::TestCases::new();
-    t.compile_fail("tests/ui/*-fail.rs");
+    t.compile_fail("tests/template/*-fail.rs");
 
-    t.pass("tests/ui/*-pass.rs");
+    t.pass("tests/template/*-pass.rs");
 }
